fixed particle rapid rerender issue when placed on each component, on navbar alone now
