+++ src/index.js
@@ -6,6 +6,7 @@ import Navbar from './Navbar';
 import ErrorPage from './ErrorPage';
 import LandingPage from './LandingPage';
 import LessonPlanner from './LessonPlanner';
+import ParticlesEffect from './ParticlesEffect';
 import {
     createBrowserRouter,
     RouterProvider,
